Bug fix and help
Fixed the hover effect for service buttons - it was an order of operations issue, I was applying the hover THEN I was applying the tooltip, and the tooltip application overrode the hover for "a reason". Put hover effect apply as the last operation and the hover effect on the service buttons works now. Also added a help dialog window/button. 

There is still a memory leak, if I toggle service status by clicking the buttons in the app, memory increase 0.1MB every time. That will be the last nut to crack...